fix admin backend

+++ controllers/admin.js
@@ -1,20 +1,34 @@
 const db_helper = require("./db_helper");
 const fs = require("fs");
 
+function handler_table(exist, is_int) {
+    // if (arg.length < 3) {return false;}
+    // return true;
+    return db_helper.existFile('./public/database/tables', exist)  || !is_int.match(/^\d+$/)
+}
+
+function create_exception(code, message) {
+    let err = new Error(message);
+    err.status = code;
+    return err;
+}
 
 exports.post_open_table = [
     function (req, res, next) { // /tables/:number/open
-        if (db_helper.getElementsFromDir('./public/database/tables').indexOf(req.params.number) === -1) {
-            let err = new Error('Error when booking a table');
-            err.status = 400;
-            return next(err);
+        if (!db_helper.existFile('./public/database/tables', req.params.number) || !req.params.number.match(/^\d+$/)) {//db_helper.getElementsFromDir('./public/database/tables').indexOf(req.params.number) === -1) {
+            // let err = new Error(`Ошибка ввода!`);
+            // err.status = 400;
+            // return next(err);
+            return next(create_exception(400, `Ошибка ввода!`))
         }
-        let path = './public/database/tables/' + req.params.number + '.json'
-        let table = JSON.parse(fs.readFileSync(path).toString())
+        let path = `./public/database/tables/${req.params.number}.json`
+        // let table = JSON.parse(fs.readFileSync(path).toString())
+        let table = db_helper.getObjectByPath(path)
         if (!table.status) {
             table.owner = ""
             table.status = true
-            fs.writeFileSync(path, JSON.stringify(table))
+            db_helper.writeObject(path, table)
+            // fs.writeFileSync(path, JSON.stringify(table))
             next()
             // res
             //     .status(200)
@@ -22,30 +36,33 @@ exports.post_open_table = [
             //     .end("The table is open successfully!");
         }
         else {
-            let err = new Error('Table is not booked!');
-            err.status = 400;
-            return next(err);
+            // let err = new Error('Столик не забронирован!');
+            // err.status = 400;
+            // return next(err);
+            return next(create_exception(400, 'Столик не забронирован!'))
         }
     },
     function (req, res, next) {
-        res.redirect('/admin/tables/' + req.params.number)
+        res.redirect(`/admin/tables/${req.params.number}`)
     }
 ]
 exports.post_add_table = [
     function (req, res, next) {
-        if (db_helper.getElementsFromDir('./public/database/tables').indexOf(req.params.number) !== -1 || req.body.number.length < 1) {
-            let err = new Error('Столик уже существует!');
-            err.status = 400;
-            return next(err);
+        if (db_helper.existFile('./public/database/tables', req.body.number)  || !req.body.number.match(/^\d+$/)) {//db_helper.getElementsFromDir('./public/database/tables').indexOf(req.params.number) !== -1 || req.body.number.length < 1) {
+            // let err = new Error('Ошибка ввода!');
+            // err.status = 400;
+            // return next(err);
+            return next(create_exception(400, `Ошибка ввода!`))
         }
-        let path = './public/database/tables/' + req.body.number + '.json'
+        let path = `./public/database/tables/${req.body.number}.json`
         let table = {
             number: req.body.number,
             status: true,
             description: req.body.description,
             owner: ""
         }
-        fs.writeFileSync(path, JSON.stringify(table))
+        db_helper.writeObject(path, table)
+        // fs.writeFileSync(path, JSON.stringify(table))
         next()
         // res
         //     .status(200)
@@ -53,17 +70,23 @@ exports.post_add_table = [
         //     .end("The table is create successfully!");
     },
     function (req, res, next) {
-        res.redirect('/admin/tables')
+        res.redirect(`/admin/tables/${req.body.number}`)
     }
 ]
 exports.post_delete_table = [
     function (req, res, next) {
-        if (db_helper.getElementsFromDir('./public/database/tables').indexOf(req.params.number) === -1) {
-            let err = new Error('Столик не существует!');
-            err.status = 400;
-            return next(err);
+        // if (db_helper.getElementsFromDir('./public/database/tables').indexOf(req.params.number) === -1) {
+        //     let err = new Error('Столик не существует!');
+        //     err.status = 400;
+        //     return next(err);
+        // }
+        if (!db_helper.existFile('./public/database/tables', req.params.number)  || !req.params.number.match(/^\d+$/)) {//db_helper.getElementsFromDir('./public/database/tables').indexOf(req.params.number) !== -1 || req.body.number.length < 1) {
+            // let err = new Error('Ошибка ввода!');
+            // err.status = 400;
+            // return next(err);
+            return next(create_exception(400, `Ошибка ввода!`))
         }
-        let path = './public/database/tables/' + req.params.number + '.json'
+        let path = `./public/database/tables/${req.params.number}.json`//'./public/database/tables/' + req.params.number + '.json'
         fs.unlinkSync(path)
         next()
         // res
@@ -78,30 +101,38 @@ exports.post_delete_table = [
 
 exports.post_add_type_menu = [
     function (req, res, next) {
-        if (db_helper.getElementsFromDir('./public/database/menu').indexOf(req.body.type) !== -1 || req.body.type.length < 1) {
-            let err = new Error('Категория уже существует!');
-            err.status = 400;
-            return next(err);
+        // if (db_helper.getElementsFromDir('./public/database/menu').indexOf(req.body.type) !== -1 || req.body.type.length < 1) {
+        if (db_helper.existFile('./public/database/menu', req.body.type) || req.body.type.length < 3) {
+            // let err = new Error('Категория уже существует!');
+            // err.status = 400;
+            // return next(err);
+            return next(create_exception(400, `Ошибка ввода!`))
         }
-        let path = './public/database/menu/' + req.body.type
+        let path = `./public/database/menu/${req.body.type}`
         fs.mkdir(path, err => {
             if (err) return next(err)
             next()
         })
+        // db_helper.createFolder(path)
+        // next()
     },
     function (req, res, next) {
-        res.redirect('/admin/menu')
+        res.redirect(`/admin/menu/${req.body.type}`)
     }
 ];
 
 exports.post_delete_type_menu = [
     function (req, res, next) {
-        if (db_helper.getElementsFromDir('./public/database/menu/' + req.params.type).length !== 0) { // FIXME: добавить проверку на существование каталога
-            let err = new Error('Не удалось удалить категорию!');
-            err.status = 400;
-            return next(err);
+        // if (db_helper.getElementsFromDir('./public/database/menu/' + req.params.type).length !== 0) { // FIXME: добавить проверку на существование каталога
+        //     let err = new Error('Не удалось удалить категорию!');
+        //     err.status = 400;
+        //     return next(err);
+        if (!db_helper.existFile('./public/database/menu/', req.params.type) ||
+                db_helper.getElementsFromDir(`./public/database/menu/${req.params.type}`).length !== 0) {
+            return next(create_exception(400, `Ошибка ввода!`))
         }
-        fs.rmdirSync('./public/database/menu/' + req.params.type)
+        // fs.rmdirSync('./public/database/menu/' + req.params.type)
+        db_helper.deleteFolder(`./public/database/menu/${req.params.type}`)
         next()
     },
     function (req, res, next) {
@@ -110,7 +141,10 @@ exports.post_delete_type_menu = [
 ]
 
 exports.get_menu_type_update = function (req, res, next) {
-    let food = db_helper.getObjectByPath('./public/database/menu/' + req.params.type + '/' + req.params.name + '.json')
+    if (!db_helper.existFile(`./public/database/menu/${req.params.type}/`, req.params.name)) {
+        return next(create_exception(400, `Ошибка ввода!`))
+    }
+    let food = db_helper.getObjectByPath(`./public/database/menu/${req.params.type}/${req.params.name}.json`)
     res.render('food_update', {
         title: 'The Krusty Krab',
         name: food.name,
@@ -123,7 +157,10 @@ exports.get_menu_type_update = function (req, res, next) {
 }
 
 exports.get_staff_type_update = function (req, res, next) {
-    let el = db_helper.getObjectByPath('./public/database/staff/' + req.params.name + '.json')
+    if (!db_helper.existFile(`./public/database/staff/`, req.params.name)) {
+        return next(create_exception(400, `Ошибка ввода!`))
+    }
+    let el = db_helper.getObjectByPath(`./public/database/staff/${req.params.name}.json`)
     res.render('food_update', {
         title: 'The Krusty Krab',
         name: el.name,
@@ -161,10 +198,11 @@ exports.get_staff_type_add = function (req, res, next) {
 
 exports.post_menu_type_delete = [
     function (req, res, next) {
-        let path = './public/database/menu/' + req.params.type + '/' + req.params.name + '.json'
-        if (db_helper.existFile('./public/database/menu/' + req.params.type, req.params.name)) {
+        let path = `./public/database/menu/${req.params.type}/${req.params.name}.json`
+        if (db_helper.existFile(`./public/database/menu/${req.params.type}`, req.params.name)) {
             let el = db_helper.getObjectByPath(path)
-            db_helper.deleteFile('./public' + el.image)
+            if (el.image !== '/images/default.webp')
+                db_helper.deleteFile(`./public${el.image}`)
             db_helper.deleteFile(path)
             next()
             // res
@@ -173,10 +211,11 @@ exports.post_menu_type_delete = [
             //     .end("The element is delete successfully!");
         }
         else
-            res
-                .status(400)
-                .contentType("text/plain")
-                .end("Удалить не удалось!");
+            return next(create_exception(400, `Ошибка ввода!`))
+            // res
+            //     .status(400)
+            //     .contentType("text/plain")
+            //     .end("Удалить не удалось!");
     },
     function (req, res, next) {
         res.redirect('/admin/menu')
@@ -185,10 +224,11 @@ exports.post_menu_type_delete = [
 
 exports.post_staff_type_delete = [
     function (req, res, next) {
-        let path = './public/database/staff/' + req.params.name + '.json'
+        let path = `./public/database/staff/${req.params.name}.json`
         if (db_helper.existFile('./public/database/staff/', req.params.name)) {
             let el = db_helper.getObjectByPath(path)
-            db_helper.deleteFile('./public' + el.image)
+            if (el.image !== '/images/default.webp')
+                db_helper.deleteFile('./public' + el.image)
             db_helper.deleteFile(path)
             next()
             // res
@@ -197,10 +237,11 @@ exports.post_staff_type_delete = [
             //     .end("The element is delete successfully!");
         }
         else
-            res
-                .status(400)
-                .contentType("text/plain")
-                .end("Удалить не удалось!");
+            return next(create_exception(400, `Ошибка ввода!`));
+            // res
+            //     .status(400)
+            //     .contentType("text/plain")
+            //     .end("Удалить не удалось!");
     },
     function (req, res, next) {
         res.redirect('/admin/staff')
@@ -209,24 +250,31 @@ exports.post_staff_type_delete = [
 
 exports.post_menu_type_update = [
     function (req, res, next) {
+        let image_path = '/images/default.webp'
+        if (req.file && req.file.filename)
+            image_path = `/images/menu/${req.file.filename}`
         let food = {
             name: req.body.name,
             description: req.body.description,
             composition: req.body.composition.split(','),
-            image: '/images/menu/' + req.file.filename
+            image: image_path//'/images/menu/' + req.file.filename
         }
-        let filedata = req.file;
-        //console.log(filedata);
-        if (!filedata){
-            let err = new Error('Ошибка загрузки изображения!');
-            err.status = 400;
-            return next(err);
-        }
+        // let filedata = req.file;
+        // //console.log(filedata);
+        // if (!filedata){
+        //     let err = new Error('Ошибка загрузки изображения!');
+        //     err.status = 400;
+        //     return next(err);
+        // }
         let path = './public/database/menu/' + req.params.type + '/' + food.name + '.json'
-        if (db_helper.existFile('./public/database/menu/' + req.params.type, food.name))
-            db_helper.deleteFile('./public' + db_helper.getObjectByPath(path).image)
+        if (db_helper.existFile('./public/database/menu/' + req.params.type, food.name)) {
+            let el = db_helper.getObjectByPath(path)
+            if (el.image !== '/images/default.webp')
+                db_helper.deleteFile('./public' + el.image)
+        }
 
-        fs.writeFileSync(path, JSON.stringify(food))
+        // fs.writeFileSync(path, JSON.stringify(food))
+        db_helper.writeObject(path, food)
         next()
         // res
         //     .status(200)
@@ -234,30 +282,37 @@ exports.post_menu_type_update = [
         //     .end("The successfully!");
     },
     function (req, res, next) {
-        res.redirect('/admin/menu/' + req.params.type)
+        res.redirect(`/admin/menu/${req.params.type}`)
     }
 ]
 
 exports.post_staff_type_update = [
     function (req, res, next) {
+        let image_path = '/images/default.webp'
+        if (req.file && req.file.filename)
+            image_path = `/images/staff/${req.file.filename}`
         let el = {
             name: req.body.name,
             post: req.body.post,
             self: req.body.self,
-            image: '/images/staff/' + req.file.filename
+            image: image_path//'/images/staff/' + req.file.filename
         }
-        let filedata = req.file;
-        //console.log(filedata);
-        if (!filedata){
-            let err = new Error('Ошибка загрузки изображения!');
-            err.status = 400;
-            return next(err);
+        // let filedata = req.file;
+        // //console.log(filedata);
+        // if (!filedata){
+        //     let err = new Error('Ошибка загрузки изображения!');
+        //     err.status = 400;
+        //     return next(err);
+        // }
+        let path = `./public/database/staff/${el.name}.json`
+        if (db_helper.existFile('./public/database/staff/', el.name)) {
+            let el_ = db_helper.getObjectByPath(path)
+            if (el_.image !== '/images/default.webp')
+                db_helper.deleteFile('./public' + el_.image)
         }
-        let path = './public/database/staff/' + el.name + '.json'
-        if (db_helper.existFile('./public/database/staff/', el.name))
-            db_helper.deleteFile('./public' + db_helper.getObjectByPath(path).image)
 
-        fs.writeFileSync(path, JSON.stringify(el))
+        // fs.writeFileSync(path, JSON.stringify(el))
+        db_helper.writeObject(path, el)
         next()
         // res
         //     .status(200)
